Allow extended grid coordinates in toLonLat/toUTM.
Implement a strict mode that rejects extended coordinates which is
necessary when searching for the closest full coordinate when evaluating
an truncated coordinate and an initial lat,lon.

Implement a relaxed mode (default) that will allow coordinates that
extend across a UTM Zone or Grid Zone boundary as long as the value
isn't ambiguous.